+++ internal/core/config.go
@@ -21,6 +21,7 @@ type Config struct {
 		Offline       bool
 		UseBinaryPath bool // Makes $SEANIME_WORKING_DIR point to the binary's directory
 		Systray       bool
+		DoHUrl        string
 	}
 	Database struct {
 		Name string